src/api/hardware_abstraction.py
"""
Hardware Abstraction Layer for Walking Robot Control Chip
Provides high-level APIs for accessing hardware capabilities
"""

import numpy as np
from typing import List, Tuple, Dict, Optional
import ctypes
import warnings

class MemoryManager:
    """Manages memory operations and prefetching for optimal performance"""

    def __init__(self):
        self.cache_config = {
            'l1_size': 32768,  # 32KB
            'l2_size': 262144, # 256KB
            'l3_size': 2097152 # 2MB
        }
        self.prefetch_queue = []
        self.attention_cache = {}

    def prefetch(self, addresses: List[int]) -> None:
        """Queue memory addresses for prefetching"""
        self.prefetch_queue.extend(addresses)

    def cache_attention_weights(self, layer_id: int, weights: np.ndarray) -> None:
        """Cache attention weights for quick access"""
        self.attention_cache[layer_id] = weights

    def get_cached_attention(self, layer_id: int) -> Optional[np.ndarray]:
        """Retrieve cached attention weights"""
        return self.attention_cache.get(layer_id)

class NPUInterface:
    """Interface for Neural Processing Unit operations"""

    def __init__(self, memory_manager: MemoryManager):
        self.memory_manager = memory_manager
        self.systolic_array_size = 8
        self.height_history = []  # Track recent height values
        self.min_step_length = 0.02  # Minimum step length when height is low
        self.target_height = 0.6  # Target operating height

    def matrix_multiply(self, a: np.ndarray, b: np.ndarray) -> np.ndarray:
        """Perform matrix multiplication using systolic array"""
        if a.shape[1] != b.shape[0]:
            raise ValueError("Matrix dimensions must match for multiplication")
        return np.dot(a, b)

    def compute_gait_pattern(self, phase: float, params: Dict) -> np.ndarray:
        """Generate walking gait pattern"""
        step_height = params.get('step_height', 0.02)  # Reduced height
        max_step_length = params.get('step_length', 0.05)  # Maximum step length
        damping = params.get('damping', 0.3)  # Strong damping for stability
        cycle_time = params.get('cycle_time', 1.0)  # Slower walking
        current_height = params.get('current_height', self.target_height)  # Get current robot height

        # Update height history and calculate trend with more samples
        self.height_history.append(current_height)
        if len(self.height_history) > 20:  # Increased history size
            self.height_history.pop(0)

        # Calculate both short-term and long-term trends
        short_window = min(5, len(self.height_history))
        long_window = min(20, len(self.height_history))

        short_trend = np.mean(np.diff(self.height_history[-short_window:])) if short_window > 1 else 0
        long_trend = np.mean(np.diff(self.height_history[-long_window:])) if long_window > 1 else 0

        # More sensitive trend detection
        height_trend = 2.0 * short_trend + long_trend

        # Initial stabilization phase (first 100 steps)
        startup_phase = min(1.0, phase / (2 * np.pi * 50))  # Gradual startup over 50 cycles

        # Height-based motion scaling with aggressive height maintenance
        target_ratio = current_height / self.target_height
        height_factor = np.clip(target_ratio ** 2, 0.3, 1.0)  # Quadratic scaling for more aggressive response
        height_error = self.target_height - current_height
        height_urgency = np.clip(1.0 + 2.0 * abs(height_error), 1.0, 3.0)

        # Adaptive step length with height-aware scaling
        trend_factor = np.clip(1.0 + 8.0 * height_trend, 0.5, 2.0)  # More aggressive trend response
        step_length = max(self.min_step_length,
                         max_step_length * height_factor * trend_factor * startup_phase)

        # Generate joint angles for walking pattern
        joint_angles = np.zeros(2)  # 2 joints: left_hip and right_hip

        # Enhanced support phase calculation
        support_phase = (1 - np.cos(phase)) / 2
        double_support = 0.5 * (1 - np.cos(2 * phase))  # Additional double support phase

        # Multi-harmonic base motion for smoother transitions
        base_motion = (0.6 * np.sin(phase) +                    # Primary forward motion
                      0.2 * np.sin(3 * phase) +                 # Triple frequency for push-off
                      0.1 * np.sin(2 * phase - np.pi/4))       # Phase-shifted double frequency

        # Height maintenance components
        height_lift = 0.3 * height_urgency * np.sin(2 * phase)  # Vertical lift component
        push_up = 0.2 * (1 - height_factor) * np.abs(np.sin(4 * phase))  # Active push-up

        # Enhanced stability components
        weight_shift = 0.15 * height_factor * np.sin(2 * phase - np.pi/6)  # Phase-shifted weight transfer
        posture_correction = 0.25 * double_support * np.sin(4 * phase)  # Enhanced during double support
        balance_term = 0.3 * support_phase * height_factor  # Height-aware balance

        # Startup stabilization
        startup_stabilizer = (1 - startup_phase) * (
            0.2 * np.sin(8 * phase) +  # High-frequency stabilization
            0.1 * np.sin(16 * phase)   # Very high-frequency fine control
        )

        # Combine all terms with adaptive damping
        stance_damping = damping * (1 + 0.5 * support_phase) * (1.5 - height_factor)
        motion_pattern = stance_damping * (
            base_motion +
            height_lift +
            push_up +
            weight_shift +
            posture_correction +
            balance_term +
            startup_stabilizer
        )

        # Apply coordinated hip motions with enhanced forward bias
        forward_bias = 0.1 * height_factor  # Add slight forward lean when height is good
        joint_angles[0] = -(step_length * motion_pattern + forward_bias)  # left hip
        joint_angles[1] = (step_length * motion_pattern + forward_bias)   # right hip

        return joint_angles

class RCUInterface:
    """Interface for Robotics Control Unit operations"""

    def __init__(self):
        self.joint_states = np.array([0.15, -0.15])  # 2 joints: left_hip and right_hip, wider initial stance
        self.joint_velocities = np.zeros(2)
        self.pid_gains = {
            'height': {'kp': 25.0, 'ki': 3.0, 'kd': 6.0},  # More aggressive height control
            'lateral': {'kp': 3.0, 'ki': 0.2, 'kd': 0.8},
            'joints': {'kp': 2.0, 'ki': 0.3, 'kd': 0.5}
        }
        self.height_setpoint = 0.6  # Desired robot height
        self.height_error_sum = 0.0
        self.last_height_error = 0.0
        self.lateral_error_sum = 0.0
        self.last_lateral_error = 0.0
        self.recovery_counter = 0  # Counter for push-up motion timing
        self.height_history = []  # Track recent height values

    def set_joint_positions(self, positions: np.ndarray) -> None:
        """Set target joint positions with stability compensation"""
        if positions.shape != (2,):
            raise ValueError("Expected 2 joint positions (left_hip, right_hip)")

        # Apply gravity compensation and stability control
        compensated_positions = self._apply_stability_control(positions)
        self.joint_states = compensated_positions

    def _apply_stability_control(self, positions: np.ndarray) -> np.ndarray:
        """Apply height and lateral stability control"""
        # Height control with weight transfer detection
        current_height = self._get_robot_height()

        # Update height history and detect rapid drops
        self.height_history.append(current_height)
        if len(self.height_history) > 20:  # Track last 20 steps
            self.height_history.pop(0)

        # Calculate both immediate and trend-based height changes
        immediate_change = self.height_history[-1] - self.height_history[-2] if len(self.height_history) > 1 else 0
        height_trend = np.mean(np.diff(self.height_history)) if len(self.height_history) > 1 else 0

        height_error = self.height_setpoint - current_height
        self.height_error_sum = min(2.0, self.height_error_sum + height_error)  # Limit integral windup
        height_derivative = height_error - self.last_height_error

        # Emergency recovery mode when height is too low or dropping rapidly
        CRITICAL_HEIGHT = 0.4
        RECOVERY_HEIGHT = 0.5  # Height at which to transition back to normal walking
        DROP_RATE_THRESHOLD = -0.0005  # More sensitive drop detection

        # Enhanced detection conditions
        in_recovery = (current_height < CRITICAL_HEIGHT or
                      height_trend < DROP_RATE_THRESHOLD or
                      immediate_change < -0.001)  # Immediate drop detection
        in_transition = CRITICAL_HEIGHT <= current_height < RECOVERY_HEIGHT

        # Detect weight transfer phase based on joint positions
        weight_transfer_factor = np.abs(positions[0] - positions[1]) / 2.0
        height_urgency = 1.0 + 4.0 * weight_transfer_factor  # Even more aggressive during transfers

        if in_recovery:
            # Emergency height recovery with enhanced push-up motion
            recovery_phase = np.sin(self.recovery_counter * 0.2)  # Faster oscillation
            self.recovery_counter += 1

            # Maximum correction during recovery
            height_urgency *= 4.0 + max(0, recovery_phase)  # Quadruple boost during upward motion
            self.height_error_sum *= 2.0  # More aggressive integral term

            # Extremely aggressive PID gains during recovery
            height_correction = height_urgency * (
                self.pid_gains['height']['kp'] * 5.0 * height_error +  # 400% stronger P term
                self.pid_gains['height']['ki'] * self.height_error_sum * 2.0 +
                self.pid_gains['height']['kd'] * height_derivative * 4.0  # Much stronger derivative
            )

            # Enhanced push-up motion pattern
            push_amplitude = 0.4 * (1 + np.abs(recovery_phase))  # Stronger push amplitude
            compensated_positions = np.array([push_amplitude, push_amplitude])  # Symmetric leg push
            compensated_positions += height_correction
            self.last_height_error = height_error
            return compensated_positions

        elif in_transition:
            # Gradual transition back to normal walking
            transition_factor = (current_height - CRITICAL_HEIGHT) / (RECOVERY_HEIGHT - CRITICAL_HEIGHT)
            self.recovery_counter = 0  # Reset recovery counter

            # More aggressive blending during transition
            height_correction = height_urgency * (
                self.pid_gains['height']['kp'] * (3.0 + 2.0 * (1 - transition_factor)) * height_error +
                self.pid_gains['height']['ki'] * self.height_error_sum * 1.5 +
                self.pid_gains['height']['kd'] * height_derivative * (3.0 + 2.0 * (1 - transition_factor))
            )

            # Gradually reintroduce lateral control
            lateral_influence = transition_factor
            compensated_positions = positions.copy() * transition_factor
            compensated_positions += height_correction
            return compensated_positions

        # Normal walking mode with preemptive height maintenance
        # Enhanced height correction with trend-based adjustment
        trend_factor = max(1.0, 2.0 - 15.0 * height_trend)  # More aggressive trend correction
        immediate_factor = max(1.0, 1.5 - 20.0 * immediate_change)  # Immediate response factor

        height_correction = height_urgency * trend_factor * immediate_factor * (
            self.pid_gains['height']['kp'] * 2.0 * height_error +  # Double base P gain
            self.pid_gains['height']['ki'] * self.height_error_sum * 1.5 +  # 50% stronger I term
            self.pid_gains['height']['kd'] * height_derivative * 2.0  # Double D term
        )
        self.last_height_error = height_error

        # Lateral stability control with height-aware adjustment
        lateral_error = -self._get_robot_lateral_position()  # Negative for correction
        self.lateral_error_sum = min(1.0, self.lateral_error_sum + lateral_error)  # Limit lateral integral
        lateral_derivative = lateral_error - self.last_lateral_error

        # Reduce lateral correction when height error is large
        height_priority = max(0.1, 1.0 - 2.0 * abs(height_error))  # More height-focused
        lateral_correction = height_priority * (
            self.pid_gains['lateral']['kp'] * lateral_error +
            self.pid_gains['lateral']['ki'] * self.lateral_error_sum +
            self.pid_gains['lateral']['kd'] * lateral_derivative
        )
        self.last_lateral_error = lateral_error

        # Apply corrections with height priority
        compensated_positions = positions.copy()
        compensated_positions += height_correction
        compensated_positions[0] += lateral_correction * 0.3  # Further reduced lateral influence
        compensated_positions[1] -= lateral_correction * 0.3

        return compensated_positions

    def _get_robot_height(self) -> float:
        """Estimate robot height from joint states"""
        LEG_LENGTH = 0.5  # Length of each leg segment
        left_angle = self.joint_states[0]
        right_angle = self.joint_states[1]
        # Use forward kinematics to calculate height
        left_height = LEG_LENGTH * np.cos(left_angle)
        right_height = LEG_LENGTH * np.cos(right_angle)
        return LEG_LENGTH + np.mean([left_height, right_height])

    def _get_robot_lateral_position(self) -> float:
        """Estimate robot lateral position from joint states"""
        LEG_LENGTH = 0.5  # Length of each leg segment
        # Calculate lateral displacement using forward kinematics
        left_lateral = LEG_LENGTH * np.sin(self.joint_states[0])
        right_lateral = LEG_LENGTH * np.sin(self.joint_states[1])
        return np.mean([left_lateral, right_lateral])

    def update_pid_gains(self, gains: dict) -> None:
        """Update PID controller gains"""
        self.pid_gains.update(gains)

    def get_joint_states(self) -> Tuple[np.ndarray, np.ndarray]:
        """Get current joint positions and velocities"""
        return self.joint_states, self.joint_velocities

class HISInterface:
    """Interface for Human Interface System operations"""

    def __init__(self):
        self.emergency_stop = False
        self.fault_status = 0x00
        self.system_temperature = 0.0

    def check_safety_status(self) -> Dict:
        """Check system safety status"""
        return {
            'emergency_stop': self.emergency_stop,
            'fault_status': self.fault_status,
            'temperature': self.system_temperature,
            'is_safe': not self.emergency_stop and self.fault_status == 0x00
        }

    def trigger_emergency_stop(self) -> None:
        """Trigger emergency stop"""
        self.emergency_stop = True
        warnings.warn("Emergency stop triggered!", RuntimeWarning)

class WalkingRobotAPI:
    """Main API for controlling the walking robot"""

    def __init__(self):
        self.memory_manager = MemoryManager()
        self.npu = NPUInterface(self.memory_manager)
        self.rcu = RCUInterface()
        self.his = HISInterface()
        self.walking_phase = 0.0
        self.initialization_steps = 0
        self.is_initialized = False
        self.gait_params = {
            'step_height': 0.1,
            'step_length': 0.2,
            'cycle_time': 1.0
        }

    def initialize(self) -> None:
        """Initialize the robot control system"""
        # Reset all subsystems
        self.walking_phase = 0.0
        self.initialization_steps = 0
        self.is_initialized = False
        self.his.emergency_stop = False
        self.rcu.joint_states = np.zeros(2)  # 2 joints: left_hip and right_hip

        # Set initial stance for stability
        initial_stance = np.array([0.1, -0.1])  # Slight outward stance
        self.rcu.set_joint_positions(initial_stance)

    def update_walking_pattern(self, dt: float) -> None:
        """Update walking pattern based on time step"""
        if not self.his.check_safety_status()['is_safe']:
            return

        current_height = self.rcu._get_robot_height()

        # Initialization phase (first 200 steps)
        if not self.is_initialized:
            self.initialization_steps += 1
            if self.initialization_steps <= 200:
                # Hold stable stance and focus on height maintenance
                stance_phase = min(1.0, self.initialization_steps / 100.0)
                stabilizing_stance = np.array([0.1, -0.1]) * stance_phase
                self.rcu.set_joint_positions(stabilizing_stance)

                # Check if height is stable
                if self.initialization_steps > 150 and abs(current_height - self.rcu.height_setpoint) < 0.05:
                    self.is_initialized = True
                return
            else:
                self.is_initialized = True  # Force initialization after 200 steps

        # Normal walking phase
        self.walking_phase += 2 * np.pi * dt / self.gait_params['cycle_time']
        self.walking_phase %= 2 * np.pi

        # Update gait parameters with current height
        gait_params = self.gait_params.copy()
        gait_params['current_height'] = current_height

        # Generate new joint positions with height awareness
        joint_positions = self.npu.compute_gait_pattern(
            self.walking_phase,
            gait_params
        )

        # Update joint controls
        self.rcu.set_joint_positions(joint_positions)

    def set_walking_parameters(self, params: Dict) -> None:
        """Set walking gait parameters"""
        self.gait_params.update(params)

    def get_system_status(self) -> Dict:
        """Get complete system status"""
        safety_status = self.his.check_safety_status()
        joint_positions, joint_velocities = self.rcu.get_joint_states()

        return {
            'safety': safety_status,
            'joint_positions': joint_positions,
            'joint_velocities': joint_velocities,
            'walking_phase': self.walking_phase,
            'gait_params': self.gait_params,
            'is_initialized': self.is_initialized,
            'initialization_steps': self.initialization_steps
        }

# Framework Integration Interfaces
class TensorFlowInterface:
    """TensorFlow integration for the walking robot hardware"""
    
    def __init__(self, robot_api: WalkingRobotAPI):
        self.robot_api = robot_api
    
    def create_custom_layer(self):
        """Create custom TensorFlow layer for hardware acceleration"""
        pass  # Implement TensorFlow custom layer

class PyTorchInterface:
    """PyTorch integration for the walking robot hardware"""
    
    def __init__(self, robot_api: WalkingRobotAPI):
        self.robot_api = robot_api
    
    def create_custom_module(self):
        """Create custom PyTorch module for hardware acceleration"""
        pass  # Implement PyTorch custom module
